Relocate pixi story stuff under components
