+++ src/components/game/dialog-manager/Text.js
