+++ src/components/game/dialog-manager/avatar/avatar.js
