+++ src/components/game/external-manager/external-manager.js
