+++ src/components/game/location-manager/location-manager.js
