+++ src/components/game/map-overlay/map-overlay.js
