+++ src/components/game/object-manager/object-manager.js
