+++ src/components/game/quest-manager/quest-manager.js
