+++ src/components/game/story-manager/story-manager.js
