+++ src/components/game/story-xml-player.js
